pyproject.toml
[tool.poetry]
name = "Vc1"
version = "1.0.0"
description = "A code that can make your discord account 24/7 on voice channels!"
authors = ["OmitGaming"]

[tool.poetry.dependencies]
python = ">=3.10,<3.15"
Flask = "^3.1.3"
requests = "^2.33.1"
websocket-client = "^1.9.0"
waitress = "^3.0.2"

[tool.poetry.dev-dependencies]
debugpy = "^1.6.2"
replit-python-lsp-server = {extras = ["yapf", "rope", "pyflakes"], version = "^1.5.9"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"